Fixup reward caching to never allow time gaps in the DB.
The previous reward caching/fetching would miss data in some cases:
1. fetch and cache 2020-01-05 to 2020-01-10
2. fetch and cache 2020-01-15 to 2020-01-20
cache now contains two discting time ranges with a gap in between

Now ask for anything that overlaps that entire gap:
* fetch 2020-01-01 to 2020-03-01
* fetch 2020-01-04 to 2020-01-16

To fix this, the DB cache is now "greedy". If hotspot earnings is
called with start and/or stop outside the current range in the DB,
we will fetch and "extend" the current DB range down to start or
up to stop or both. This leads to sometimes fetching more data than
needed for the current work, but ensures the DB has no time gaps.